third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_boundary_events_before_pointerrawupdate.https-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer boundary events should be fired before `pointerrawupdate`
  assert_equals: expected "pointerover, pointerenter, pointerrawupdate, pointermove" but got "pointerrawupdate, pointerover, pointerenter, pointermove"
Harness: the test ran to completion.
